chore: cleanup commented code and some logging

+++ dash-sdk-java/src/main/cpp/java_list_value.cpp
@@ -60,12 +60,12 @@ jobject fermented_vec_to_java_list_Value(JNIEnv * jenv, Vec_platform_value_Value
     jclass valueClass = jenv->FindClass("org/dashj/platform/sdk/PlatformValue");
     jmethodID valueConstructor = jenv->GetMethodID(valueClass, "<init>", "(JZ)V");
     //printf("PlatformValue: %lx, %lx\n", valueClass, valueConstructor);
-    printf("processing list input = 0x%lx, count: %d\n", input, input->count);
+    printf("processing list input = 0x%lx, count: %d\n", (unsigned long) input, (int) input->count);
     for (uintptr_t i = 0; i < input->count; ++i) {
-        printf("%d, %lx, %lx\n", i, input, input->values);
+        printf("%d, %lx, %lx\n", (int) i, (unsigned long) input, (unsigned long)input->values);
         jobject value = jenv->NewObject(valueClass, valueConstructor,
                                         (jlong) input->values[i], false);
-        printf("new object: %lx", value);
+        printf("new object: %lx", (unsigned long)value);
         jenv->CallObjectMethod(hashMapInstance, addMethod, value);
     }
